+++ applet-version/README.md
@@ -1,0 +1,6 @@
+
+Build:
+`mvn clean install`
+
+Run:
+`appletviewer target/classes/index.html`
