+++ applet-version/src/main/resources/images/black_bishop.gif
