+++ applet-version/src/main/resources/images/black_king.bmp
