+++ applet-version/src/main/resources/images/black_king.gif
