+++ applet-version/src/main/resources/images/black_knight.gif
