+++ applet-version/src/main/resources/images/black_pawn.gif
