+++ applet-version/src/main/resources/images/black_queen.bmp
