+++ applet-version/src/main/resources/images/black_queen.gif
