+++ applet-version/src/main/resources/images/black_rook.gif
