+++ applet-version/src/main/resources/images/white_bishop.gif
