+++ applet-version/src/main/resources/images/white_king.gif
