+++ applet-version/src/main/resources/images/white_knight.bmp
