+++ applet-version/src/main/resources/images/white_pawn.gif
