+++ applet-version/src/main/resources/images/white_queen.bmp
