+++ applet-version/src/main/resources/images/white_rook.gif
