+++ applet-version/src/main/resources/index.html
@@ -11,7 +11,7 @@
 
 <body>
 <center>
-<applet code="Chess.java" height="256" width= "256"></applet><br>
+<applet code="com.uberchess.applet.Chess" height="256" width= "256"></applet><br>
 Thomas Pensyl's UberChess v. 0.5<br><br>
 <p>To start a new game refresh this page.</p><br>
 <p>If the applet is grey, then your browser doesn't have java set up right, all I can say is to either install java or try it on a different computer or a different browser. It should work on the school computers in Internet Explorer.</p>
