--- run.bat
@@ -1,1 +1,0 @@
-appletviewer index.html
